+++ inc/core/post-types.php
@@ -43,10 +43,7 @@ function extrachill_docs_register_post_type() {
 		'show_ui'            => true,
 		'show_in_menu'       => true,
 		'query_var'          => true,
-		'rewrite'            => array(
-			'slug'       => 'doc',
-			'with_front' => false,
-		),
+		'rewrite'            => false,
 		'capability_type'    => 'post',
 		'has_archive'        => true,
 		'hierarchical'       => true,